feat: Add preload attribute to Video component in Reel
The Video component in Video.tsx has been updated to include the preload attribute with a value of "auto". This change improves the video loading performance by preloading the video when it is in view. The addition of the preload attribute enhances the user experience and optimizes the overall performance of the Reel component.