feat: Add NewCategory component for creating new categories
A new component called NewCategory has been added to allow users to create new categories. This component provides a form for users to input a category name and select an image from a list of GIFs. The selected category is then added to the database. This addition enhances the user experience and improves the organization of video content.