refactor: Update profile page UI and add copyright information
This commit updates the UI of the profile page by modifying the styles of the tabs and the logout button. It also adds a copyright information section at the bottom of the page, displaying the year and the creator's Twitter handle. These changes enhance the visual appeal and provide necessary information to the users.