refactor: Add canPlay state to Video component
The code changes in the components/Reel/Video.tsx file add a new state variable, canPlay, to the Video component. This state variable is used to control the playback of the video when the status is set to "playing". The useEffect hook has been updated to include the canPlay state as a dependency, ensuring that the video only plays when it is ready. This change improves the functionality and performance of the Video component.